Add trace.go
